# aspnetcore/src/IO.Swagger/Controllers/AdminsApi.cs
/*
 * Simple Temperature API
 *
 * This API is meant as an exchange mechanism for temperature readings
 *
 * OpenAPI spec version: 1.0.0
 * Contact: stephen@solenoidsoft.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class AdminsApiController : ControllerBase
    { 
        /// <summary>
        /// adds a temperature sensor reading
        /// </summary>
        /// <remarks>Adds a reading</remarks>
        /// <param name="body">Inventory item to add</param>
        /// <response code="201">item created</response>
        /// <response code="400">invalid input, object invalid</response>
        /// <response code="409">an existing item already exists</response>
        [HttpPost]
        [Route("/Solenoid/Temperature/1.0.0/temperature")]
        [ValidateModelState]
        [SwaggerOperation("AddReading")]
        public virtual IActionResult AddReading([FromBody]TemperatureReading body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409);

            throw new NotImplementedException();
        }
    }
}